examples/python-fastapi/requirements.txt
# ==========================================
# FastAPI 本番用依存関係（2025年版）
# ==========================================

# FastAPI本体
fastapi[standard]==0.115.0

# ASGIサーバー（本番用）
uvicorn[standard]==0.32.0

# データベース
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.3

# Pydantic（バリデーション・シリアライゼーション）
pydantic==2.9.2
pydantic-settings==2.6.0

# 認証
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Redis（キャッシュ・セッション）
redis==5.2.0

# CORS
python-dotenv==1.0.1

# HTTP クライアント（外部API連携用）
httpx==0.27.2

# ロギング
python-json-logger==3.1.0
